+++ test/pc/setup-dkml-darwin_arm64.sh
@@ -2438,7 +2438,7 @@ do_install_compiler() {
     section_begin "install-compiler-$do_install_compiler_NAME" "Install OCaml compiler for $do_install_compiler_NAME switch"
     opamrun pin list --switch "$do_install_compiler_NAME"
     # shellcheck disable=SC2086
-    opamrun upgrade --switch "$do_install_compiler_NAME" --yes dkml-base-compiler conf-dkml-cross-toolchain ${ocaml_options:-}
+    opamrun upgrade --switch "$do_install_compiler_NAME" --yes dkml-base-compiler conf-dkml-cross-toolchain "dkml-host-abi-${dkml_host_abi}" "dkml-target-abi-${dkml_target_abi}" ${ocaml_options:-}
     section_end "install-compiler-$do_install_compiler_NAME"
 }
 do_summary() {